consistently convert db_recno_t values to u_long for snprintf
When using the bundled db.1.85, db_recno_t is set to u_int32_t,
which may not equivalent to long.
Since a long length modifier is used in the format,
explicitly convert the value to long.
This was already being done in a couple of places, but not consistently.

+++ vi/vs_refresh.c
@@ -822,7 +822,8 @@ vs_modeline(SCR *sp)
 	if (O_ISSET(sp, O_RULER)) {
 		vs_column(sp, &curcol);
 		len =
-		    snprintf(buf, sizeof(buf), "%lu,%lu", sp->lno, curcol + 1);
+		    snprintf(buf, sizeof(buf), "%lu,%lu",
+			(u_long)sp->lno, curcol + 1);
 
 		midpoint = (cols - ((len + 1) / 2)) / 2;
 		if (curlen < midpoint) {
